+++ example/lib/page/home_page.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:oktoast/oktoast.dart';
 import 'package:photo_manager/photo_manager.dart';
 import 'package:photo_manager_example/widget/nav_button.dart';
 import 'package:provider/provider.dart';
@@ -12,7 +13,7 @@ import 'filter_option_page.dart';
 import 'gallery_list_page.dart';
 
 class NewHomePage extends StatefulWidget {
-  const NewHomePage({Key? key}) : super(key: key);
+  const NewHomePage({super.key});
 
   @override
   State<NewHomePage> createState() => _NewHomePageState();
@@ -45,7 +46,7 @@ class _NewHomePageState extends State<NewHomePage> {
               title: 'Get all gallery list',
               onPressed: _scanGalleryList,
             ),
-            if (Platform.isIOS)
+            if (Platform.isIOS || Platform.isAndroid)
               CustomButton(
                 title: 'Change limited photos with PhotosUI',
                 onPressed: _changeLimitPhotos,
@@ -66,6 +67,7 @@ class _NewHomePageState extends State<NewHomePage> {
             _buildPngCheck(),
             _buildNotifyCheck(),
             _buildFilterOption(watchProvider),
+            if (Platform.isIOS || Platform.isMacOS) _buildPathFilterOption(),
           ],
         ),
       ),
@@ -116,6 +118,18 @@ class _NewHomePageState extends State<NewHomePage> {
   }
 
   Future<void> _scanGalleryList() async {
+    final permissionResult = await PhotoManager.requestPermissionExtend(
+        requestOption: PermissionRequestOption(
+      androidPermission: AndroidPermission(
+        type: readProvider.type,
+        mediaLocation: true,
+      ),
+    ));
+    if (!permissionResult.hasAccess) {
+      showToast('no permission');
+      return;
+    }
+
     await readProvider.refreshGalleryList();
     if (!mounted) {
       return;
@@ -223,6 +237,18 @@ class _NewHomePageState extends State<NewHomePage> {
         Navigator.push<void>(
           context,
           MaterialPageRoute<void>(builder: (_) => const FilterOptionPage()),
+        );
+      },
+    );
+  }
+
+  Widget _buildPathFilterOption() {
+    return CustomButton(
+      title: 'Change path filter options.',
+      onPressed: () {
+        Navigator.push<void>(
+          context,
+          MaterialPageRoute<void>(builder: (_) => const DarwinPathFilterPage()),
         );
       },
     );