+++ example/lib/widget/video_widget.dart
@@ -8,10 +8,10 @@ import '../util/log.dart';
 
 class VideoWidget extends StatefulWidget {
   const VideoWidget({
-    Key? key,
+    super.key,
     required this.entity,
     this.usingMediaUrl = true,
-  }) : super(key: key);
+  });
 
   final AssetEntity entity;
   final bool usingMediaUrl;
@@ -81,7 +81,7 @@ class _VideoWidgetState extends State<VideoWidget> {
       if (!mounted || url == null) {
         return;
       }
-      _controller = VideoPlayerController.network(url)
+      _controller = VideoPlayerController.networkUrl(Uri.parse(url))
         ..initialize()
         ..addListener(() => setState(() {}));
       setState(() {});