+++ ios/Classes/core/PMManager.h
@@ -12,6 +12,7 @@ typedef void (^ChangeIds)(NSArray<NSString *> *);
 @class PMFilterOption;
 @class PMFilterOptionGroup;
 @class PMThumbLoadOption;
+@class PMPathFilterOption;
 
 #import "PMProgressHandlerProtocol.h"
 #import "PMResultHandler.h"
@@ -28,13 +29,9 @@ typedef void (^AssetResult)(PMAssetEntity *);
 
 @property(nonatomic, strong) NSObject <PMConvertProtocol> *converter;
 
-- (BOOL)isAuth;
-
 + (void)openSetting:(NSObject<PMResultHandler>*)result;
 
-- (void)setAuth:(BOOL)auth;
-
-- (NSArray<PMAssetPathEntity *> *)getAssetPathList:(int)type hasAll:(BOOL)hasAll onlyAll:(BOOL)onlyAll option:(NSObject<PMBaseFilter> *)option;
+- (NSArray<PMAssetPathEntity *> *)getAssetPathList:(int)type hasAll:(BOOL)hasAll onlyAll:(BOOL)onlyAll option:(NSObject <PMBaseFilter> *)option pathFilterOption:(PMPathFilterOption *)pathFilterOption;
 
 - (NSUInteger)getAssetCountFromPath:(NSString *)id type:(int)type filterOption:(NSObject<PMBaseFilter> *)filterOption;