+++ lib/src/types/entity.dart
@@ -10,8 +10,8 @@ import 'package:flutter/rendering.dart';
 
 import '../filter/base_filter.dart';
 import '../filter/classical/filter_option_group.dart';
+import '../filter/path_filter.dart';
 import '../internal/constants.dart';
-import '../internal/editor.dart';
 import '../internal/enums.dart';
 import '../internal/plugin.dart';
 import '../internal/progress_handler.dart';
@@ -27,13 +27,13 @@ class AssetPathEntity {
   AssetPathEntity({
     required this.id,
     required this.name,
-    @Deprecated('Use assetCountAsync instead. This will be removed in 3.0.0')
-        this.assetCount = 0,
     this.albumType = 1,
     this.lastModified,
     this.type = RequestType.common,
     this.isAll = false,
     PMFilter? filterOption,
+    this.darwinSubtype,
+    this.darwinType,
   }) : filterOption = filterOption ??= FilterOptionGroup();
 
   /// Obtain an entity from ID.
@@ -66,13 +66,6 @@ class AssetPathEntity {
   ///  * iOS/macOS: Album/Folder name.
   final String name;
 
-  /// Total assets count of the album.
-  ///
-  /// The synchronized count will cause performance regression on iOS,
-  /// here the asynchronized getter [assetCountAsync] is preferred.
-  @Deprecated('Use assetCountAsync instead. This will be removed in 3.0.0')
-  final int assetCount;
-
   /// Total assets count of the path with the asynchronized getter.
   Future<int> get assetCountAsync => plugin.getAssetCountFromPath(this);
 
@@ -101,6 +94,16 @@ class AssetPathEntity {
 
   /// The collection of filter options of the album.
   final PMFilter filterOption;
+
+  /// The darwin collection type, in android, the value is always null.
+  ///
+  /// If the [albumType] is 2, the value will be null.
+  final PMDarwinAssetCollectionType? darwinType;
+
+  /// The darwin collection subtype, in android, the value is always null.
+  ///
+  /// If the [albumType] is 2, the value will be null.
+  final PMDarwinAssetCollectionSubtype? darwinSubtype;
 
   /// Call this method to obtain new path entity.
   static Future<AssetPathEntity> obtainPathFromProperties({
@@ -143,7 +146,7 @@ class AssetPathEntity {
       return ConvertUtils.convertToPathList(
         result.cast<String, dynamic>(),
         type: type,
-        optionGroup: optionGroup,
+        filterOption: optionGroup,
       ).first;
     }
     throw error;
@@ -252,7 +255,7 @@ class AssetPathEntity {
       return ConvertUtils.convertToPathList(
         result.cast<String, dynamic>(),
         type: type,
-        optionGroup: filterOptionGroup ?? filterOption,
+        filterOption: filterOptionGroup ?? filterOption,
       ).first;
     }
     return null;
@@ -266,6 +269,8 @@ class AssetPathEntity {
     RequestType? type,
     bool? isAll,
     PMFilter? filterOption,
+    PMDarwinAssetCollectionType? darwinType,
+    PMDarwinAssetCollectionSubtype? darwinSubtype,
   }) {
     return AssetPathEntity(
       id: id ?? this.id,
@@ -275,6 +280,8 @@ class AssetPathEntity {
       type: type ?? this.type,
       isAll: isAll ?? this.isAll,
       filterOption: filterOption ?? this.filterOption,
+      darwinSubtype: darwinSubtype ?? this.darwinSubtype,
+      darwinType: darwinType ?? this.darwinType,
     );
   }