+++ lib/src/utils/convert_utils.dart
@@ -4,6 +4,7 @@
 
 import '../filter/base_filter.dart';
 import '../filter/classical/filter_option_group.dart';
+import '../filter/path_filter.dart';
 import '../types/entity.dart';
 import '../types/types.dart';
 
@@ -13,7 +14,7 @@ class ConvertUtils {
   static List<AssetPathEntity> convertToPathList(
     Map<String, dynamic> data, {
     required RequestType type,
-    PMFilter? optionGroup,
+    PMFilter? filterOption,
   }) {
     final List<AssetPathEntity> result = <AssetPathEntity>[];
     final List<Map<dynamic, dynamic>> list =
@@ -27,7 +28,7 @@ class ConvertUtils {
         convertMapToPath(
           item.cast<String, dynamic>(),
           type: type,
-          optionGroup: optionGroup ?? FilterOptionGroup(),
+          filterOption: filterOption ?? FilterOptionGroup(),
         ),
       );
     }
@@ -47,7 +48,7 @@ class ConvertUtils {
   static AssetPathEntity convertMapToPath(
     Map<String, dynamic> data, {
     required RequestType type,
-    PMFilter? optionGroup,
+    PMFilter? filterOption,
   }) {
     final int? modified = data['modified'] as int?;
     final DateTime? lastModified = modified != null
@@ -57,12 +58,17 @@ class ConvertUtils {
       id: data['id'] as String,
       name: data['name'] as String,
       // ignore: deprecated_member_use_from_same_package
-      assetCount: data['assetCount'] as int? ?? 0,
       albumType: data['albumType'] as int? ?? 1,
-      filterOption: optionGroup ?? FilterOptionGroup(),
+      filterOption: filterOption ?? FilterOptionGroup(),
       lastModified: lastModified,
       type: type,
       isAll: data['isAll'] as bool,
+      darwinType: PMDarwinAssetCollectionTypeExt.fromValue(
+        data['darwinAssetCollectionType'],
+      ),
+      darwinSubtype: PMDarwinAssetCollectionSubtypeExt.fromValue(
+        data['darwinAssetCollectionSubtype'],
+      ),
     );
     return result;
   }